+++ lib/PaymentTipAndPayDine/PaymentTipAndPayDiPresenter.dart
@@ -124,7 +124,7 @@ class PayBillCheckoutPresenter extends PayBillCheckoutContaractor {
   void payBillCheckOut(
     int restId,
     // String currency,
-    double amount,
+    int amount,
     BuildContext context,
   ) {
     ApiBaseHelper().post<PaycheckoutNetbanking>(
